+++ src/theme/components/checkboxGroup.ts
@@ -2,7 +2,12 @@ import { checkboxRadioGroupStyles } from "./global";
 
 const CheckboxGroup = {
   parts: ["helper", "stack"],
-  baseStyle: checkboxRadioGroupStyles,
+  baseStyle: ({ isFullWidth = false }) => ({
+    ...checkboxRadioGroupStyles,
+    stack: {
+      width: isFullWidth ? "100%" : checkboxRadioGroupStyles.stack.width,
+    },
+  }),
 };
 
 export default CheckboxGroup;