+++ CHANGELOG.md
@@ -8,6 +8,14 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 
 ## Prerelease
 
+### Adds
+
+- Adds the `isFullWidth` prop to the `CheckboxGroup` component. This sets the checkbox's wrapper element to be full width for labels that need to span its container.
+
+### Changes
+
+- Updates the `Checkbox`'s `labelText` prop to accept strings and JSX Elements.
+
 ## 0.25.12 (March 18, 2022)
 
 ## Adds